+++ examples/e2e.py
@@ -12,7 +12,11 @@ memory_strategy = 'high'  # can be high or low depending on memory availability/
 product_dst_dir = Path('../notebooks/los-angeles')  # directory to save the final products
 apply_water_mask = True  # apply water mask to the data
 n_lookbacks = 3  # number of lookbacks to use for change confirmation within SAS
-water_mask_path = Path('../notebooks/los-angeles/water_mask.tif')  # path to an existing water mask file
+water_mask_path = None  # Path('../notebooks/los-angeles/water_mask.tif')  # path to an existing water mask file
+device = 'cpu'  # can be cpu, cuda, mps or best
+n_workers_for_norm_param_estimation = 4  # number of workers for normal parameter estimation
+batch_size_for_despeckling = 50  # batch size for despeckling
+n_workers_for_despeckling = 1  # number of workers for despeckling
 
 
 # Run the workflow
@@ -27,4 +31,8 @@ run_dist_s1_workflow(
     apply_water_mask=apply_water_mask,
     n_lookbacks=n_lookbacks,
     water_mask_path=water_mask_path,
+    device=device,
+    n_workers_for_norm_param_estimation=n_workers_for_norm_param_estimation,
+    batch_size_for_despeckling=batch_size_for_despeckling,
+    n_workers_for_despeckling=n_workers_for_despeckling,
 )